+++ test/domain/ac_date_range_test.dart
@@ -1,4 +1,4 @@
-import 'package:appcraft_ui_calendar_flutter/src/domain/ac_date_range.dart';
+import 'package:appcraft_ui_calendar_flutter/src/domain/src/ac_date_range.dart';
 import 'package:test/test.dart';
 
 void main() {