+++ test/widgets/ac_time_input_widget_test.dart
@@ -6,7 +6,7 @@ void main() {
   Widget buildTestApp(Widget child) {
     return MaterialApp(
       theme: ThemeData(
-        extensions: [ACCalendarThemeData()],
+        extensions: [ACCalendarThemeExtension(data: ACCalendarThemeData())],
       ),
       home: Scaffold(body: child),
     );